+++ br/pkg/storage/BUILD.bazel
@@ -66,7 +66,6 @@ go_library(
         "@org_golang_google_api//iterator",
         "@org_golang_google_api//option",
         "@org_golang_x_oauth2//google",
-        "@org_golang_x_sync//errgroup",
         "@org_uber_go_atomic//:atomic",
         "@org_uber_go_zap//:zap",
     ],
